fix: T05b.03-Exercise-PolishAsyncTask: Solve the lack of loading problem
You were assigning a value to mGithubJson when that was not necessary.  Leave
it blank.  Check for null then depending on that either deluver result or
forceLoad.  When you create the variable with a value, it messes up the loading.
Look at onStartLoading in this URL:
https://medium.com/google-developers/making-loading-data-on-android-lifecycle-aware-897e12760832
The value is NOT initialized.