feat: Create the actual Database
Except that in real life you would have an update stmt in onUpdate that did not
drop the table.  Currently, you drop the table and then recreate it.
For real, you should do a command to change the table structure; while keeping
all the data the same.  After that you should increment the version number.
Remember, if you change the schema you also have to change the Contract.