feat: Add vector graphic to project
To add a .svg file to the project rClick on res folder.
Choose New > Vector Asset
Hit Local asset radio button
Type the pathh to the asset in the path box, followed by a slash,
followed by its whole name and press enter.  If it is not an icon
hit override radio button so it is not square shaped.
Name it art_, or ic_, or something [name].svg
Hit next then enter.